fix(test): address PR review — compose lifecycle, sweep gaps, doc snippets
Review feedback from cubic, Greptile, Copilot and CodeScene. Several were
real defects in code added by this PR.

Compose lifecycle was broken (Greptile P1, cubic P2)
----------------------------------------------------
`--env-file tests/.env` was passed only to `test-integration-up`, but the
compose file declares LOCALSTACK_AUTH_TOKEN with a `${...:?}` guard that
compose evaluates on EVERY parse. For a developer whose token lives only in
tests/.env — the documented path — `make test-integration-down` could not
read the compose file at all, so a started stack could never be stopped. The
`ps` probe in `test-integration` hit the same error but swallowed it with
`2>/dev/null`, so it always reported zero running services and restarted the
stack unnecessarily.

Introduced a single COMPOSE_TEST variable used by up/down/ps. Verified:
`down` now removes containers with the token only in tests/.env.

The token guard also only checked that tests/.env EXISTS (cubic P3), so an
empty or fully commented file bypassed the actionable error and fell through
to the raw compose `:?` message. It now greps for a non-empty assignment;
verified against both an empty file and a commented-out key.

Sweep guards had real gaps
--------------------------
- test_cli_help_markup.py missed dashed identifiers: Rich swallows
  `[vault-sync]` exactly like `[vault]`, but the pattern excluded `-`, so a
  dashed TOML section could vanish from --help with the sweep still green
  (cubic P2). Added `-`, plus meta-tests asserting the detector sees every
  documented section-name shape and that the prescribed `\[` escape really
  renders a bare bracket.
- While adding those, the ad-hoc `startswith("env_file")` metavar filter
  proved wrong for `[paths]`. Click metavars are now derived from the app's
  real parameter names instead of guessed by prefix.
- overridesCompat.test.ts never descended into scoped packages, so a nested
  minimatch under `@scope/pkg/node_modules/` would have been skipped by the
  very guard that exists to catch it (cubic P3). Now recurses through `@*`.

Documentation
-------------
- The guard-command spec snippets called `_load_constants()` without
  defining it, so the examples were non-runnable (cubic P2 x2). Both blocks
  now include the helper, copied from the real scanner/gitleaks.py.
- The integration spec's canonical Container Setup example still omitted the
  token, so copying it produced a container that dies at boot with the exact
  error the surrounding prose warns about (cubic P3).
- The manual SOPS install wrote straight to /usr/local/bin/sops, so an
  interrupted download could truncate an existing binary (cubic P3). Now
  stages to a temp file on the same filesystem and moves it into place.
- CLI docs: Arguments tables still used the old uppercase names while the
  synopses show what typer 0.27 actually renders (Copilot x2, cubic x2).
  Verified against live `--help` that the synopses are correct — `guard`
  really does render `[paths]`, not `[PATHS]` — so the tables were aligned
  to the real metavars rather than reverting the accurate synopses.

Code health (CodeScene)
-----------------------
- `_all_help_strings` flagged as Bumpy Road; split into small helpers
  (_command_help / _param_help / _param_names) with no nested conditionals.
- test_infisical.py exceeded the 600-line threshold; the CLI-migration
  regression classes moved to tests/scanner/test_infisical_cli_migration.py.

Verified: ruff check + format clean, pyrefly 0 errors, bandit 0 issues,
markdownlint clean, workflow YAML parses, vscode 79 tests, and the
non-integration suite 3864 passed / 8 skipped.
