+++ docs/cli/guard.md
@@ -35,7 +35,7 @@ paths are provided, the current directory is scanned.
 
 | Argument | Description | Default |
 | :-- | :-- | :-- |
-| `PATHS` | Files or directories to scan | `.` |
+| `[paths]` | Files or directories to scan | `.` |
 
 ## Options
